% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_all.R
\name{de_all}
\alias{de_all}
\title{Differential Expression analysis using all combination}
\usage{
de_all(dat, idx)
}
\arguments{
\item{dat}{dataframe: RNA-seq count table.  row: samples, column: genes}

\item{idx}{factor: group of data. E.g. factor(c(1,1,2,2)); factor(c("A","A","B","B"))}
}
\description{
Description
}
\examples{
#
# data(fpkm2)
# index <-
}
