% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_temlate.R
\name{dir_template}
\alias{dir_template}
\title{Create directory for RNA-seq pipeline}
\usage{
dir_template(prj)
}
\arguments{
\item{prj}{character: path of a project name}
}
\description{
Create Project directories for RNA-seq pipeline using hisat2-stringtie.
   Then fastq files move to './project/fastq'　directory, followning execution of 'rskoseq::rep_hisat2', for read mapping using hisat2.
}
\examples{
#
# project_name <- "~/pub/dat/sampledata/rnaseq/project1"
# dir_template(prj = project_name)
# system(paste0("tree ", project_name))
# project1
# ├── command_log.txt
# ├── fastq
# ├── qa
# ├── res_hisat2
# │   └── failalign
# └── res_stringtie
#     ├── ballgown
#     ├── gff
#     ├── mgff
#     └── tab
# 9 directories, 1 file

}
