% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpkm2.R
\docType{data}
\name{fpkm2}
\alias{fpkm2}
\title{fpkm2}
\format{An object of dataframe with with 72 rows and 1004 columns.}
\usage{
data(fpkm2)
}
\description{
The FPKM dataframe from RNA-seq. The first four columns is the categolical information of 72 samples.
   The rest of 1000 columns is FPKM of genes.
}
\examples{
data(fpkm2)
fpkm2[ 1:6, 1:6 ]
}
\keyword{datasets}
