% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqfilter.R
\name{fqfilter}
\alias{fqfilter}
\title{Filtering and trimming of fastq files using ShortRead package}
\usage{
fqfilter (fqd, sffx, outd, qs, nhead, ntail, droplen)
}
\arguments{
\item{fqd}{character: the fully path of a directory of fastq files, or the path of fastq files.}

\item{sffx}{character: suffix of fastq files. The default value is ".fastq.gz"}

\item{outd}{output directory}

\item{qs}{ASCII character corresponding to q-score.}

\item{nhead}{integer: Remove leading nucleotides.}

\item{ntail}{integer: Remove trailing nucleotides.}

\item{droplen}{integer: drop length}
}
\description{
filtering and trimming of fastq files and output filtered and compressed fastq files
}
\examples{
\dontrun{
# output filterd and compressed fastq files ##
fqd <- system.file("extdata/E-MTAB-1147", package = "ShortRead")
sffx <- ".fastq.gz"
outd <- "~/pub/sampledata/rnaseq/project1/qcfq"
unlink(outd, recursive = T)
fqfilter(fqd=fqd, outd=outd, droplen = 30)

# filtered fastq
qcfq <- list.files(outd, "qc.fastq.gz", full.names =T)
}
}
