% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_read.R
\name{gc_read}
\alias{gc_read}
\title{Quality assesment with ShortRead package and return ggplot object.}
\usage{
gc_read(fqdir, suffix, prefix, facet_col)
}
\arguments{
\item{fqdir}{A vector of file path of fastq files, or dir path, containing fastq files}

\item{suffix}{A pattern of fastq file suffix. The default is ".fastq.gz"}

\item{prefix}{A vector of samples name. The default values are names of fastq files containing in fqdir, which substitute 'suffix' character.}

\item{facet_col}{facet of sequence content and quality score per sample plot.}
}
\value{
The list of ggplot objects returns and write.table for quality assesment with 'qa'.
}
\description{
This functions returns ggplot objects and create data.frame of GC content per read
}
\examples{
## arguments
# p <- system.file("extdata/E-MTAB-1147", package = "ShortRead")
# sffx <- ".fastq.gz"
# fqs <- list.files(p, sffx)
# prfx <- sapply(strsplit(fqs, "/"), function(x)sub(sffx, "", tail(x, n=1)))

## execution
# res_qa <- gc_read(fqdir=p, suffix=sffx, prefix=prfx, facet_col=2)
# do.call(gridExtra::grid.arrange, c(res_qa, list(ncol=2)))

}
