% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_summary.R
\name{genome_summary}
\alias{genome_summary}
\title{Summary of genomic sequences from fasta file}
\usage{
genome_summary(in_f, N, genome)
}
\arguments{
\item{in_f}{The input file path with fasta format, or DNAStringSet object from Biostrings package.}

\item{N}{numeric vector: If calcuration for N90(L90), N50(L50), and N10(L10), N = c(90, 50, 10)}

\item{genome}{numeric: genome size. The default is NULL. The genome size is calculated from the sum of the widths of the input fasta files.}
}
\description{
Summary of genomic sequences from fasta file
}
\examples{
# fas <- system.file("extdata/L_casei_bl23.ASM2648v1.dna.toplevel.fa.gz", package = "rskodat")
# genome_summary(in_f = fas, N=c(90,50,10))
}
