% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_qa.R
\name{gg_qa}
\alias{gg_qa}
\title{Quality assesment with ShortRead package and return ggplot object.}
\usage{
gg_qa(filepath, suffix, prefix, facet_col)
}
\arguments{
\item{filepath}{A vector of file path of fastq files, or dir path, containing fastq files}

\item{suffix}{A pattern of fastq file suffix. eg. ".fastq.gz"}

\item{prefix}{A vector of samples name(optional).}

\item{facet_col}{facet}
}
\value{
list of data frame for quality assesment with 'qa'
    and list ofggplot objects
}
\description{
This functions returns ggplot objects
}
\examples{
## arguments
# p <- system.file("extdata/E-MTAB-1147", package = "ShortRead")
# sffx <- ".fastq.gz"
# fls <- list.files(p, sffx, full.names = TRUE)
# prfx <- sapply(strsplit(fls, "/"), function(x)sub(sffx, "", tail(x, n=1)))
## execution
# res_qa <- gg_qa(filepath=fls, suffix=sffx, prefix=prfx, facet_col=2)
# do.call(gridExtra::grid.arrange, c(res_qa, list(ncol=2)))

}
