% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_cnt.R
\name{kmer_cnt}
\alias{kmer_cnt}
\title{Counting kmer of nucleotide sequenc}
\usage{
kmer_cnt(filepath, format, k, n, per_site)
}
\arguments{
\item{filepath}{Input file path}

\item{format}{Choose "fastq" or "fasta" format}

\item{k}{Length of kmer}

\item{n}{Random sampling numbers for ShortRead::FastqSampler}

\item{per_site}{logical: kmer count per site of all reads of fastq, or not}
}
\value{
kmer table or kmer count at position
}
\description{
kmer counting from fastq or fasta format file and return kmer table
}
\examples{
# fastq which reads are same
fq <- list.files(system.file("extdata/E-MTAB-1147", package = "ShortRead"), full.names = TRUE)
kmer_dat1 <- kmer_cnt(filepath = fq[1], format = "fastq", k = 7, per_site = TRUE)
# kmer_dat2 <- kmer_cnt(filepath = fq[1], format = "fastq", k = 7, per_site = FALSE)
}
