% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_strm.R
\name{kmer_strm}
\alias{kmer_strm}
\title{Counting kmer of nucleotide sequence}
\usage{
kmer_strm(x, k, n, resfmt)
}
\arguments{
\item{x}{An input fastq file path or DNAStrinset object.}

\item{k}{integer: Length of kmer}

\item{n}{integer: chunk size for splitting fastq data. The default value is 1L for no splitting.}

\item{resfmt}{character: "total" as total k-mer, or "pos" as k-mer contents at read position.}
}
\value{
kmer table or kmer count at position
}
\description{
kmer counting from fastq or fasta format file and return kmer table
}
\examples{
\dontrun{
# fastq which reads are same length
fq <- list.files(system.file("extdata/E-MTAB-1147", package = "ShortRead"), full.names = TRUE)

# k-mer table
kmer_tab <- rskoseq::kmer_strm(fq[1], 7)

# k-mer table with parallel processing
kmer_tab <- rskoseq::kmer_strm(fq[1], 7, 5000)

# k-mer contents at read position
kmer_content <- rskoseq::kmer_strm(fq[1], 7, 5000, "pos")

# A DNAStringset object as input which reads are different length
fa <- ShortRead::readFastq(fq[1]) \%>\%
  ShortRead::trimTails(., k = 2, a = "4", halfwidth = 2) \%>\%
  ShortRead::sread()
res <- rskoseq::kmer_strm(fa, 7, 5000)
}
}
