% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplot.R
\name{maplot}
\alias{maplot}
\title{maplot}
\usage{
maplot(dat, fdr, gp1, gp2)
}
\arguments{
\item{dat}{dataframe: result of DEG analysis using 'tmm_cnt'}

\item{fdr}{q-value default 0.05}

\item{gp1, gp2}{group name}
}
\description{
maplot
}
