% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maprate.R
\name{maprate}
\alias{maprate}
\title{Parse hisat2 log file}
\usage{
maprate(fp, lab)
}
\arguments{
\item{fp}{file path of bowtie log}

\item{lab}{file name as sample name}
}
\value{
data frame of mapping rate and ggplot object
}
\description{
Parse hisat2 log file and extract mapping rate of all samples.
}
\examples{
log <- system.file("extdata", "hisat2_log.txt", package = "rskoseq")
label <- c("A","B","C","D")
res <- maprate(fp = log, lab = label)
}
