% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawfqext.R
\name{rawfqext}
\alias{rawfqext}
\title{rawfqext}
\usage{
rawfqext(rawfq_dir, prefix)
}
\arguments{
\item{rawfq_dir}{fastq containing directory}

\item{prefix}{sample name}
}
\description{
merge NEXTSEQ output fastq, which created by BaseSpace
}
\examples{
# rawfq_dir <- "~/pub/sampledata/NEXTSEQ_OUT"
# prefix <- c("S1","S2")
# rawfqext(rawfq_dir, prefix)
}
