% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_hisat2.R
\name{rep_hisat2}
\alias{rep_hisat2}
\title{Replicate execution of read alignment using hisat2}
\usage{
rep_hisat2(alndir, idx, project, fqdir, paired, suffix_fq, ...)
}
\arguments{
\item{alndir}{character: the name of alignment directory. results output}

\item{idx}{character: the fully path of hisat2 index name.}

\item{project}{logical:default is TRUE. If does not create project directory using rskoseq::project_rnsq, it is FALSE}

\item{fqdir}{character: the fully path of fastq files. The default is 'paste0(dirname(alndir), "/fastq")'.
If this directory containing still analyzed fastq and additional fastq files,}

\item{paired}{logical: paired or single read. If paired end, the names of fastq file must be "_R1" and "_R2".}

\item{suffix_fq}{character: suffix of fastq files. The default is ".fastq.gz"}

\item{...}{additional hisat2 options. E.g.  "--no-spliced-alignment"}
}
\description{
The NGS read alignment using hisat2 for multiple samples.
    The input and output directory must be created by 'rskoseq::project_rnsq'.
}
\examples{
## project TRUE (all result created at directories predetermined by 'project_rnsq')
# alndir <- "~/pub/sampledata/rnaseq/project1/test.h2"
# idx <- "~/db/index/hisat2_idx/CriGri_1.0.ens"
# rep_hisat2(alndir = alndir, idx = idx, paired=F)
## paired ----
# alndir <- "~/pub/sampledata/rnaseq/project1/test2.h2"
# fqd <- "~/pub/sampledata/rnaseq/project1/paired_fastq"
# rep_hisat2(alndir = alndir, idx = idx, fqdir = fqd, suffix_fq = "_sub.fastq.gz", paired=T)
#
## project FALSE (all result create under the alignment directory)
# alndir <- "~/pub/sampledata/rnaseq/project1/
# rep_hisat2(prjd = prj, idx = idx, paired=FALSE, alnd = "alignment2")

}
