% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revcomp.R
\name{revcomp}
\alias{revcomp}
\title{Reverse and/or Complement of Nucleotide sequence as character string.}
\usage{
revcomp(x, strand)
}
\arguments{
\item{x}{character string of nucleotide sequence}

\item{strand}{character: 'rev','comp', or 'revcomp'}
}
\description{
Reverse and/or Complement of Nucleotide sequence as character string.
}
\examples{
seq <- "GCATNGCR"
revcomp(seq, "rev")
revcomp(seq, "comp")
revcomp(seq, "revcomp")
}
