% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_norm.R
\name{tcc_norm}
\alias{tcc_norm}
\title{Normarization of RNA-seq count data using tmm}
\usage{
tcc_norm(dat, column, gp, method)
}
\arguments{
\item{dat}{A data frame, matrix. The 'dat' contains columns which names samples and rows contains genes}

\item{column}{Count data columns without id column. The default is 1:ncol(dat)}

\item{gp}{replicate group}

\item{method}{a pipe line select from   # 1:'DEGES/TbT', 2:'DEGES/edgeR', 3:'iDEGES/edgeR', 4:'DEGES/DESeq',
5:'iDEGES/DESeq', 6:'iDEGES/DESeq2'.}
}
\value{
ggobject which containing result of 'TCC::getNormalizedData' ,
}
\description{
This functions returns the normarized count data of multiple groups
}
\examples{
\dontrun{
# sample data of rna-seq
fpkm_rep3 <- rskodat::fpkm
fpkm_rep2 <- rskodat::fpkm[c(1,2,4,5)]
fpkm_norep <- rskodat::fpkm[c(1, 4)]

# arguments
gp1 <- rep(1:4, each=3)
gp2 <- rep(1:2, each=2)
gp3 <- c(1,2)

# normalize
# if you choose method 4,5, or 6, require the 'TCC' library load.
tbt_nfpkm <- rskoseq::tcc_norm(dat = fpkm_rep2, column = 1:ncol(fpkm_rep2), gp = gp2, method = 1)
edgr_nfpkm <- rskoseq::tcc_norm(dat = fpkm_rep2, column = 1:ncol(fpkm_rep2), gp = gp2, method = 2)
iedgr_nfpkm <- rskoseq::tcc_norm(dat = fpkm_rep2, column = 1:ncol(fpkm_rep2), gp = gp2, method = 3)
dsq_nfpkm <- rskoseq::tcc_norm(dat = fpkm_norep, column = 1:ncol(fpkm_norep), gp = gp3, method = 4)
idsq_nfpkm <- rskoseq::tcc_norm(dat = fpkm_norep, column = 1:ncol(fpkm_norep), gp = gp3, method = 5)

TCC::plot.TCC(TCC::estimateDE(nfpkm1$tcc, test.method = "edger", FDR = 0.05))
TCC::plot.TCC(TCC::estimateDE(nfpkm2$tcc, test.method = "edger", FDR = 0.05))
TCC::plot.TCC(TCC::estimateDE(nfpkm3$tcc, test.method = "deseq2", FDR = 0.05))
}

}
