% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmm_norm.R
\name{tmm_norm}
\alias{tmm_norm}
\title{Normarization of RNA-seq count data using tmm}
\usage{
tmm_norm(dat, column, gp, method)
}
\arguments{
\item{dat}{A data frame, matrix. The 'dat' contains columns which names genes and rows contains samples.}

\item{column}{Count data columns without id column. The default is 1:ncol(dat)}

\item{gp}{replicate group}

\item{method}{a pipe line select from 1:"iDEGES/TbT", 2:"iDEGES/edgeR", 3:"iDEGES/DESeq".}
}
\value{
ggobject which containing resulto of 'TCC::getNormalizedData' ,
}
\description{
This functions returns the normarized count data of multi group
}
\examples{
# # sample data of rna-seq
# fpkm <- rskodat::fpkm
# gp <- sort(as.integer(factor(sapply(strsplit(names(fpkm), "_"),
#           function(x) paste(head(x,2), collapse = "")))))
# nfpkm <- rskoseq::tmm_norm(dat = fpkm, column = 1:ncol(fpkm), gp = gp, method = 2)
# fpkm[1:4,1:4]; nfpkm[1:4,1:4]

}
