% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degall.R
\name{degall}
\alias{degall}
\title{Differential Expression analysis using all combination}
\usage{
degall(dat, idx, normalize)
}
\arguments{
\item{dat}{dataframe: RNA-seq count table.  row: samples, column: genes}

\item{idx}{factor: group of data. E.g. factor(c(1,1,2,2)); factor(c("A","A","B","B"))}

\item{normalize}{integer: If dat is already normalized count data, 'norm' is NULL. The default value is NULL
If dat is not normalized, select a pipe line number from 1:"DEGES/TbT", 2:"DEGES/edgeR", 3:"DEGES/DESeq".}
}
\value{
ggobject which containing result of 'TCC::estimateDE', without normalized count data.
}
\description{
split count data with index and Differential expression analysis of all comination groups.
}
\examples{
## normalized fpkm
# nfpkm <- rskodat::nfpkm[c(1:3, 10:12, 19:21, 28:30),]
# index <- factor(paste0(nfpkm$runs, ".", nfpkm$days),
#                 levels = unique(paste0(nfpkm$runs, ".", nfpkm$days)))
# res <- rskoseq::degall(dat=nfpkm[-1:-4], idx=index, normalize=NULL)
## get result of estimateDE
# head(res$data)
## ggplot
# ggplus::facet_multiple(res, facets="comp", ncol = 3, nrow = 5)

}
