% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degall.R
\name{degall}
\alias{degall}
\title{Differential Expression analysis using all combination}
\usage{
degall(dat, idx, normalized, meth_norm, param_fdr)
}
\arguments{
\item{dat}{dataframe: RNA-seq count table.  row: samples, column: genes}

\item{idx}{factor: group of data. E.g. factor(c(1,1,2,2)); factor(c('A','A','B','B'))}

\item{normalized}{logical: The 'dat' is a normalized count data or not.}

\item{meth_norm}{integer: Choose from the six pipe line numbers below 1:'DEGES/TbT', 2:'DEGES/edgeR', 3:'iDEGES/edgeR',
4: 'DEGES/DESeq', 5: iDEGES/DESeq, 6: iDEGES/DESeq2. The default value is 2}

\item{param_fdr}{numeric: fdr value.}
}
\value{
ggplot object which containing result of 'TCC::estimateDE', without normalized count data.
}
\description{
split count data with index and Differential expression analysis of all comination groups.
}
\examples{
\dontrun{
# sample data of rna-seq
fpkm_rep3 <- rskodat::fpkm
fpkm_rep2 <- rskodat::fpkm[c(1,2,4,5)]
fpkm_norep <- rskodat::fpkm[c(1,4)]

# aruguments
gp <- sapply(strsplit(names(fpkm_rep3), "_"), function(x) paste(x[1:2], collapse = "_"))
index0 <- factor(rep(1:4, each = 3))
index1 <- factor(gp, levels=unique(gp))
index2 <-factor(rep(1:2, each=2))
index3 <- factor(c(1, 2))

# two-group with replicate
# If you choose 4,5,or 6 which using DESeq, you should load library TCC.
res1 <- rskoseq::degall(dat = fpkm_rep2, idx = index2, meth_norm = 1)
res2 <- rskoseq::degall(dat = fpkm_rep2, idx = index2, meth_norm = 2)
res3 <- rskoseq::degall(dat = fpkm_rep2, idx = index2, meth_norm = 3)
res4 <- rskoseq::degall(dat = fpkm_rep2, idx = index2, meth_norm = 4)
res5 <- rskoseq::degall(dat = fpkm_rep2, idx = index2, meth_norm = 5)
res6 <- rskoseq::degall(dat = fpkm_rep2, idx = index2, meth_norm = 6)

# multi-group with replicate
res3m <- rskoseq::degall(dat = fpkm_rep3, idx = index0, meth_norm = 3)
res6m <- rskoseq::degall(dat = fpkm_rep3, idx = index1, meth_norm = 6)

# two samples without replicate
res3s <- rskoseq::degall(dat = fpkm_norep, idx = index3, meth_norm = 3)
res4s <- rskoseq::degall(dat = fpkm_norep, idx = index3, meth_norm = 4)
print(res3s$maplot)

# get result of estimateDE
head(res3m$deg)

# redraw MA-plot
library(dplyr)
deg_dat <- res3m$deg \%>\% filter(comp \%in\% levels(.$comp)[1:3])
num_de <- res3m$num_deg \%>\% filter(comp \%in\% levels(.$comp)[1:3])

ggplot2::ggplot(deg_dat, ggplot2::aes(x = a.value, y = m.value, colour = fct)) +
  ggplot2::geom_point(size = 0.3) +
  ggplot2::scale_color_manual(values =
    c("non-DEG"="#BEBEBE80", "up"="#FF000080", "down" ="#0000FF80")) +
  ggplot2::theme_minimal(base_size = 15) +
  ggplot2::labs(x = "A=(log2(G2)+log2(G1))/2", y = "M=log2(G2)-log2(G1)", colour = "") +
  ggplot2::theme(legend.position = "top") +
  ggplot2::guides(colour = ggplot2::guide_legend(override.aes = list(alpha=1, size = 5))) +
  ggplot2::geom_text(data = add_de, ggplot2::aes(x = x, y = y, label = value, colour = key),
    size = 5, show.legend = F) +
  ggplot2::facet_wrap(~comp, ncol = 3)
}

}
