% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_temlate.R
\name{dir_template}
\alias{dir_template}
\title{Create directory for RNA-seq pipeline}
\usage{
dir_template(prj)
}
\arguments{
\item{prj}{character: path of a project name}
}
\description{
Create directory for RNA-seq pipeline using hisat2-stringtie
}
\examples{
#
dir_template(prj = "~/pub/dat/sampledata/rnseq/project1")

}
