% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_read.R
\name{gc_read}
\alias{gc_read}
\title{GC content per read}
\usage{
gc_read(fqdir, outdir, suffix, prefix, facet_col)
}
\arguments{
\item{fqdir}{character: The file path of fastq files, or dir path, containing fastq files}

\item{outdir}{output directory}

\item{suffix}{A pattern of fastq file suffix. The default is ".fastq.gz"}

\item{prefix}{A vector of samples name. The default values are names of fastq files containing in fqdir, which substitute 'suffix' character.}

\item{facet_col}{facet of sequence content and quality score per sample plot.}
}
\value{
The list of ggplot objects returns and write.table for quality assesment with 'qa'.
}
\description{
This functions returns ggplot objects of GC content per read
}
\examples{
\dontrun{
# arguments
fqdir <- system.file("extdata/E-MTAB-1147", package = "ShortRead")
suffix <- ".fastq.gz"
outdir <- "~/pub/sampledata/rnaseq/project1/qa"

# execution
res_qa <- rskoseq::gc_read(fqdir, outdir, suffix, facet_col=2)
do.call(gridExtra::grid.arrange, c(res_qa, list(ncol=2)))

}
}
