% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_summary.R
\name{genome_summary}
\alias{genome_summary}
\title{Summary of genomic sequences from fasta file}
\usage{
genome_summary(in_f, N, genome)
}
\arguments{
\item{in_f}{The input file path with fasta format, or DNAStringSet object from Biostrings package.}

\item{N}{numeric vector: If calcuration for N90(L90), N50(L50), and N10(L10), N = c(90, 50, 10)
The N50 length is defined as the shortest sequence length at 50 percent of the genome. L50 is the number of contigs whose summed length is N50.}

\item{genome}{numeric: genome size. The default is NULL. The genome size is calculated from the sum of the widths of the input fasta files.}
}
\description{
Summary of genomic sequences from fasta file
}
\examples{
\dontrun{
fas <- "~/db/genome/CHOK1GS_HDv1/CHOK1GS_HDv1.dna.toplevel.fa.gz"
res <- genome_summary(in_f = fas, N = c(90,50,10))

# summary of genomic sequences
res$summary

# base frequency
res$base_freq

# disribution of contig lengs
hist(log10(res$dist_cntg))

}
}
