% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_qa.R
\name{gg_qa}
\alias{gg_qa}
\title{Quality assesment with ShortRead package and return ggplot object.}
\usage{
gg_qa(fqdir, suffix, prefix, qareport, facet_col, outdir, ow)
}
\arguments{
\item{fqdir}{A vector of file path of fastq files, or dir path, containing fastq files}

\item{suffix}{A pattern of fastq file suffix. The default is ".fastq.gz"}

\item{prefix}{A vector of samples name. The default values are names of fastq files containing in fqdir, which substitute 'suffix' character.}

\item{qareport}{logical. reporting or not. The default value is FALSE.}

\item{facet_col}{integer. facet of sequence content and quality score per sample plot.}

\item{outdir}{output directory. The default values is same at fqdir.}

\item{ow}{logical. It must be TRUE if overwritten. The default value is FALSE.}
}
\value{
The list of ggplot objects returns and write.table for quality assesment with 'qa'.
}
\description{
This functions returns ggplot objects and create report of qa. The qa report is created at above of the fastq directory.
}
\examples{
\dontrun{
# arguments
p <- system.file("extdata/E-MTAB-1147", package = "ShortRead")
sffx <- ".fastq.gz"
prfx <- sub(sffx,"",list.files(p, sffx))

# execution
res_qa <- rskoseq::gg_qa(fqdir=p, suffix=sffx, prefix=prfx, facet_col=2)
do.call(gridExtra::grid.arrange, c(res_qa, list(ncol=2)))
}
}
