% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggblast.R
\name{ggblast}
\alias{ggblast}
\title{Simple visualization of local blast search output}
\usage{
ggblast(blast_out, drawtype)
}
\arguments{
\item{blast_out}{dataframe: result of rskoseq::rblast output}

\item{drawtype}{character number: 1: facet_wrap with query id, 2: A panel per query, 3: All in one panel}
}
\description{
The result of blast output using rsko::rblast is dataframe, or
    local blast output dat with outfmt '6'. The data has these columns, at least, "qseqid", "sseqid", "pident", "qstart", "qend", "sstart", "send", "qlen", "slen","sstrand".
}
\examples{
\dontrun{
## sample fasta of Biostrings package
fp <- system.file("extdata", "AtMlos.fna", package="rskodat")
bndb <- "~/db/cdna/TAIR10_cdna" # set you environment
bnout <- rskoseq::rblast(fp, "-", "blastn", bndb, "-num_threads 4")
res1 <- ggblast(bnout, 1)
res2 <- ggblast(bnout, 2)
res3 <- ggblast(bnout, 3)
}
}
