% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maprate.R
\name{maprate}
\alias{maprate}
\title{Parse hisat2 or bowtie2 log file}
\usage{
maprate(fp, lab, pair, algnr)
}
\arguments{
\item{fp}{file path of bowtie log}

\item{lab}{character: file name as sample name.}

\item{pair}{logical: paired end read or nod. The default value is TRUE.}

\item{algnr}{character: select an alignr from 'hisat2' or 'bowtie2'. The default value is "bowtie2".}
}
\value{
data frame of mapping rate and ggplot object
}
\description{
Parse hisat2 log file and extract mapping rate of all samples.
}
\examples{
h2.paired.log <- system.file("extdata", "hisat2_log.txt", package = "rskoseq")
label <- c("A","B","C","D")
res1 <- maprate(fp = h2.paired.log, lab = label)
res2 <- maprate(fp = h2.paired.log, pair = T, lab = label, algnr = "hisat2")
}
