% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_rnsq.R
\name{project_rnsq}
\alias{project_rnsq}
\title{Create directory for RNA-seq pipeline}
\usage{
project_rnsq(prjd, alnd, algnr)
}
\arguments{
\item{prjd}{character: path of a project directory name}

\item{alnd}{character: path of a alignment directory name, the default is "alignment1"}

\item{algnr}{character: select an alignment program from "hisat2"(default), "rsem", "others"}
}
\description{
Create Project directories for RNA-seq pipeline using hisat2-stringtie.
   Then fastq files move to './project/fastq'　directory, followning execution of 'rskoseq::rep_hisat2' or 'rskoseq::rep_rsem', for read mapping.
}
\examples{
\dontrun{
# create new project
prj <- "~/pub/sampledata/rnaseq/project1"
project_rnsq(prjd = prj, alnd = "test.h2", algnr = "hisat2")
project_rnsq(prjd = prj, alnd = "test.rsm", algnr = "rsem")

# now there is a project directory, create another alignment directory.
system(paste("tree", prj))

}
}
