% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawfqext.R
\name{rawfqext}
\alias{rawfqext}
\title{rawfqext}
\usage{
rawfqext(rawfq_dir, paired, prefix)
}
\arguments{
\item{rawfq_dir}{fastq containing directory}

\item{paired}{logical, if TRUE is paired end read.}

\item{prefix}{Characters of sample name, The default value is NULL, collect from directories name.}
}
\description{
merge NEXTSEQ output fastq, which created by BaseSpace
}
\examples{
# rawfq_dir <- "~/pub/sampledata/fastq/NEXTSEQ_OUT"
# prefix <- c("S1","S2")
# rawfqext(rawfq_dir, prefix)
}
