% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawfqext.R
\name{rawfqext}
\alias{rawfqext}
\title{rawfqext}
\usage{
rawfqext(rawfq_dir, paired, prefix)
}
\arguments{
\item{rawfq_dir}{fastq containing directory}

\item{paired}{logical, if TRUE is paired end read.}

\item{prefix}{Characters of sample name, The default value is NULL, collect from directories name.}
}
\description{
Merging fastq files of NEXTSEQ output, lane splitted, created by BaseSpace
}
\examples{
\dontrun{
rawfq_dir <- "~/pub/sampledata/fastq/NEXTSEQ"
file.remove(list.files(rawfq_dir, "*.gz", full.names = T))
prefix <- c("3AD","3AE")
rawfqext(rawfq_dir, prefix)
}
}
