% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rblast.R
\name{rblast}
\alias{rblast}
\title{Local blast search with outfmt 6}
\usage{
rblast(in_f, out_f, program, db, ...)
}
\arguments{
\item{in_f}{character: input file path of multifasta format.}

\item{out_f}{character: output file path or stdout as "-".}

\item{program}{character: select a blast search program "blastn", "blastp", "blastx", "tblastn", "tblastx"}

\item{db}{blast database name or file path of fasta: if you gives fasta file path, 'makeblastdb' was executed.
it is not file path. if multiple database using, corresponding out put files path gave as 'out_f'.}

\item{...}{additional parameter as character strings E.g. "-num_threads 4 -task megablast".}
}
\value{
blast output of 'outfmt 6' and several other column returnd as named data frame.
 if all query "No hits found", could not create data.frame
}
\description{
This functions returns a data frame as BLAST output, if multiple databese gave as an arugment, they are all in one,
    which fomat is outfmt 6 and additional column. The blast outformat is enforced as follows, '-outfmt "6 std qlen slen sstrand salltitles"'.
}
\examples{
\dontrun{
## create blast data base
# com0 <- c("makeblastdb -in TAIR10_chr_all.fas -out TAIR10 -dbtype nucl -parse_seqids")
# system(com0, intern = T)

# sample fasta of rsko package
in_fna <- system.file("extdata", "AtMlos.fna", package="rskodat")
bndb <-  "~/db/cdna/TAIR10_cdna"
bnout <- rblast(in_f = in_fna, out_f = "-", program = "blastn", db = bndb, "-num_threads 4")

# multiple db
db1 <- "~/db/cdna/TAIR10_cdna"
db2 <- "~/db/cdna/kegg_plant"
dbs <- c(db1,db2)
in_fna <- system.file("extdata", "AtMlos.fna", package="rskodat")
bnouts <- rblast(in_f = in_fna, out_f = "-", program = "blastn", db = dbs, "-num_threads 4")
}
}
