% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_hisat2.R
\name{rep_hisat2}
\alias{rep_hisat2}
\title{Replicate execution of read mapping using hisat2}
\usage{
rep_hisat2(prj, fqdir, suffix_fq, idx, paired, ...)
}
\arguments{
\item{prj}{character: project directory. create by 'rskoseq::dir_template'}

\item{fqdir}{character: path of fastq files. default: paste0(prj, "/fastq")}

\item{suffix_fq}{character: suffix of fastq files. The default is ".fastq.gz"}

\item{idx}{character: hisat2 index name}

\item{paired}{logical: paired or sinle read}

\item{...}{additional hisat2 options. E.g.  "--no-spliced-alignment"}
}
\description{
Description
}
\examples{
#
# # arguments ----
# prj <- "~/pub/dat/sampledata/rnaseq/project1"
# # fqdir <- paste0(prj, "/fastq")
# # suffix_fq <- ".fastq.gz"
# idx <- "~/db/index/hisat2_idx/SD0218_11_a_contig01"
# paired <- FALSE
# # excution ----
# rep_hisat2(prj=prj, idx=idx, paired=TRUE, ...="--no-spliced-alignment")
}
