% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_rsem.R
\name{rep_rsem}
\alias{rep_rsem}
\title{Continuous execution of sequence alignment using RSEM}
\usage{
rep_rsem (path_aln, idx_name, suffix_fq)
}
\arguments{
\item{path_aln}{alignment directory path, created by rskoseq::project_rnsq.}

\item{idx_name}{rsem index file path}

\item{suffix_fq}{suffix of fastq files. The default value is ".fastq.gz"}
}
\description{
Description
}
\examples{
##
### rep rsem
# idx <- "~/db/index/rsem_idx/cge25207.add"
# alnd <- "~/pub/sampledata/rnaseq/project1/rsem1/"
# prj <- "~/pub/sampledata/rnaseq/project1"
# rep_rsem(path_prj = prj, alnd = alnd, idx_name=idx)
}
