% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_rsem.R
\name{rep_rsem}
\alias{rep_rsem}
\title{Continuous execution of sequence alignment using RSEM}
\usage{
rep_rsem (alndir, fqdir, paired, idx_name, suffix_fq, prefix_fq, ...)
}
\arguments{
\item{alndir}{character: output alignment directory path, created by rskoseq::project_rnsq.}

\item{fqdir}{character: the fully path of fastq files. The default is 'paste0(dirname(alndir), "/fastq")'.
If this directory containing still analyzed fastq and additional fastq files,}

\item{paired}{logical: paired or single read. If paired end, the names of fastq file must be "_R1" and "_R2".}

\item{idx_name}{rsem index file path}

\item{suffix_fq}{suffix of fastq files. The default value is ".fastq.gz"}

\item{prefix_fq}{character: The default value is `sub(suffix_fq,"", list.files(fqdir))`}

\item{...}{additional rsem-calcurate-expression options. E.g.  "--fragment-length-max"}
}
\description{
Description
}
\examples{
\dontrun{
# create index
rsem-prepare-reference --bowtie2 ref.fasta refname
rsem-prepare-reference --bowtie2 --transcript-to-gene-map list.txt ref.fasta refname

# single
idx <- "~/db/index/rsem_idx/cge25207.add"
alnd <- "~/pub/sampledata/rnaseq/project1/res_single_rsm"
unlink(alnd, recursive = T)
rskoseq::project_rnsq("~/pub/sampledata/rnaseq/project1", "res_single_rsm", "rsem")
rskoseq::rep_rsem(alndir=alnd, idx_name=idx)

# single (no project -> you must specifie fastq directory)
alnd <- "~/pub/sampledata/rnaseq/res_rsm"
fqd <- "~/pub/sampledata/rnaseq/project1/fastq"
idx <- "~/db/index/rsem_idx/cge25207.add"
unlink(alnd, recursive = T)
rskoseq::rep_rsem(alndir=alnd, idx_name=idx, fqdir = fqd, suffix_fq = "_R1.fastq.gz")

# paired
alnd <- "~/pub/sampledata/rnaseq/project1/res_paired_rsm"
fqd <- "~/pub/sampledata/rnaseq/project1/paired_fastq"
idx <- "~/db/index/rsem_idx/cge25207.add"
unlink(alnd, recursive = T)
rskoseq::project_rnsq("~/pub/sampledata/rnaseq/project1", "res_paired_rsm", "rsem")
rskoseq::rep_rsem(alndir = alnd, fqdir = fqd, paired = T, idx_name = idx)

}
}
