% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_stringtie.R
\name{rep_stringtie}
\alias{rep_stringtie}
\title{Consecutive execution of Transcript assembly and quantification for RNA-Seq using stringtie}
\usage{
rep_stringtie(bamdir, suffix_bam, guide_gff, res_dir, ...)
}
\arguments{
\item{bamdir}{character: the name of alignment directory, sorted-bam files searched from under this directory.
If designate your own sorted-bam containing directory, give the path of the directory
and 'res_dir' must be directory name under the 'bamdir'.}

\item{suffix_bam}{character: The default is ".sort.bam".}

\item{guide_gff}{The file path of guide gff.}

\item{res_dir}{output directory path, the default is 'paste0(dirname(bamdir), "/res_stringtie")'}

\item{...}{additional options of stringtie. E.g. "-e"}
}
\description{
Consecutive processiong of stringtie for multiple samples, then FPKM and cov data table are created.
   Execution following from 'rskoseq::project_rnsq', 'rskoseq::rep_hisat2'.
}
\examples{
\dontrun{
# create project directory
alnd <- "~/pub/sampledata/rnaseq/project1/test1.h2"
idx <- "~/db/index/hisat2_idx/CriGri_1.0.ens.add"
unlink(alnd, recursive = T)
rskoseq::project_rnsq("~/pub/sampledata/rnaseq/project1", "test1.h2", "hisat2")
system(paste("tree", alnd))

# run rep_hisat2
rskoseq::rep_hisat2(alndir = alnd, idx = idx, paired = F)
system(paste("tree", alnd))

# run rep_stringtie
bamdir <- "~/pub/sampledata/rnaseq/project1/test1.h2/res_hisat2"
guide <- "~/db/index/hisat2_idx/CriGri_1.0.ens.add.gff"
rskoseq::rep_stringtie(bamdir = bamdir, guide_gff = guide)
system(paste("tree", alnd))
}
}
