% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmm_norm.R
\name{tmm_norm}
\alias{tmm_norm}
\title{Normarization of RNA-seq count data using tmm}
\usage{
tmm_norm(dat, column, gp, method)
}
\arguments{
\item{dat}{A data frame, matrix. The 'dat' contains columns which names genes and rows contains samples.}

\item{column}{Count data columns without id column. The default is 1:ncol(dat)}

\item{gp}{replicate group}

\item{method}{a pipe line select from 1:"DEGES/TbT", 2:"DEGES/edgeR", 3:"DEGES/DESeq".}
}
\value{
dataframe of normalized count data
}
\description{
This functions returns the normarized count data of multi group
}
\examples{
# sample data of rna-seq
data(fpkm2)
## normalized count
# group <- fpkm2$reps
# nfpkm <- tmm_norm(dat = fpkm2, column = 5:ncol(fpkm2), gp = group, method = 2)
}
